+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
  - All arguments are now correctly passed to winetricks.
+ - Games that haven't been launched at least once are now excluded properly.
 
 ## [1.1.1] - 2019-01-20
 ### Added